refactor: extracts hook to mirror the width of a node
