+++ packages/palette/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./useClickOutside"
 export * from "./useIsomorphicLayoutEffect"
 export * from "./usePosition"
 export * from "./useUpdateEffect"
+export * from "./useWidthOf"
